Remove unneeded VOLUME directive
This option causes Docker to create a volume every time this container
is run.  However, since this container does not write anything to the
volume that isn't regenerated quickly every time it is run, the
directive serves no purpose.

By itself that isn't a biug problem.  However, once the VOLUME directive
has been added to a location, no files can be added to it, whether it be
in the same image or any image that inherits from it.  Worse still,
inheriting images cannot remove the VOLUME -- they're stuck with it.

Removing it doesn't effect the normal use cases of using --volume on the CLI or
volume bind mounts in a Docker Compose file.

What it does allow is for images that inherit from this one to add vpn
configuration to /vpn with no shenanigans.